refactor(env): Extract `page` into separate file
